5.7. Ипотека [475130]

+++ src/ru/job4j/loop/Mortgage.java
@@ -2,11 +2,11 @@ package ru.job4j.loop;
 
 public class Mortgage {
     public static int year(double amount, double salary, double percent) {
-        double year = 0;
+        int year = 0;
         while (amount > 0) {
             year++;
             amount = (amount + amount * percent / 100) - salary;
         }
-        return (int) year;
+        return year;
     }
 }
